I init <OwnerWalletContents /> component
