+++ components/ProfilePage/PreWalletCollection.tsx
@@ -1,35 +1,22 @@
-import { useEffect, useState } from "react"
+import { useState } from "react"
 import { useMediaQuery } from "usehooks-ts"
-import { useRouter } from "next/router"
 import Media from "../../shared/Media"
 import UnlockModal from "./UnlockModal"
 import TrainModal from "./TrainModal"
 import { useProfileProvider } from "../../providers/ProfileContext"
-import getProfileFormattedCollection from "../../lib/getProfileFormattedCollection"
 import SmartWalletContents from "./SmartWalletContents"
+import OwnerWalletContents from "./OwnerWalletContents"
 
 const PreWalletCollection = () => {
   const isMobile = useMediaQuery("(max-width: 1024px)")
-  const router = useRouter()
-  const { expandedMore, setExpandedMore, isEditable } = useProfileProvider()
-  const { address } = router.query as any
-  const [ownedNfts, setOwnedNfts] = useState([])
+  const { expandedMore, setExpandedMore } = useProfileProvider()
   const [openUnlockModal, setOpenUnlockModal] = useState(false)
   const [openTraninModal, setOpenTrainModal] = useState(false)
-
-  useEffect(() => {
-    const init = async () => {
-      const response = await getProfileFormattedCollection(address)
-
-      // Set the state
-      setOwnedNfts(response)
-    }
-    init()
-  }, [address])
 
   const toggleUnlockModal = () => {
     setOpenUnlockModal(!openUnlockModal)
   }
+
   const toggleTraninModal = () => {
     setOpenTrainModal(!openTraninModal)
   }
@@ -117,58 +104,10 @@ const PreWalletCollection = () => {
               />
             </button>
           </div>
-          <div
-            className="grid grid-cols-3 xs:grid-cols-4 lg:grid-cols-6 gap-x-[3px] lg:gap-x-[15px] gap-y-[5px] pt-[15px] mt-[20px]
-                  h-[140px] lg:h-[287px] 
-                  overflow-y-auto overflow-x-hidden
-                  pr-2"
-          >
-            {ownedNfts.map((data, i) => (
-              // eslint-disable-next-line react/no-array-index-key
-              <div key={i} className="flex flex-col items-center gap-y-[5px]">
-                {data.image ? (
-                  <img
-                    className="w-[30px] h-[30px] samsungS8:w-[35px] samsungS8:h-[35px] lg:w-[93px] lg:h-[93px] bg-white rounded-[5px] lg:rounded-[15px]"
-                    src={data.image}
-                    alt={data.label}
-                  />
-                ) : (
-                  <div className="w-[30px] h-[30px] samsungS8:w-[35px] samsungS8:h-[35px] lg:w-[93px] lg:h-[93px] bg-white rounded-[5px] lg:rounded-[15px]" />
-                )}{" "}
-                <div
-                  className="text-[6px] samsungS8:text-[7px] xs:text-[8px] lg:text-[12px] font-quicksand font-bold text-white
-                              w-[30px] samsungS8:w-[40px] lg:!w-[90px] text-center
-                              flex flex-col items-center gap-y-[2px]"
-                >
-                  <div className="w-full">{data.label}</div>
-                  {isEditable && (
-                    <div>
-                      {data.type === "cre8or" &&
-                        (data.isLocked ? (
-                          <button type="button" onClick={() => setOpenUnlockModal(true)}>
-                            <Media
-                              type="image"
-                              containerClasses="w-[13.54px] h-[16.83px]"
-                              link="/assets/Profile/locked.svg"
-                              blurLink="/assets/Profile/locked.png"
-                            />
-                          </button>
-                        ) : (
-                          <button onClick={() => setOpenTrainModal(true)} type="button">
-                            <Media
-                              type="image"
-                              containerClasses="w-[14.8px] h-[17px]"
-                              link="/assets/Profile/unlocked.svg"
-                              blurLink="/assets/Profile/unlocked.png"
-                            />
-                          </button>
-                        ))}
-                    </div>
-                  )}
-                </div>
-              </div>
-            ))}
-          </div>
+          <OwnerWalletContents
+            setOpenUnlockModal={setOpenUnlockModal}
+            setOpenTrainModal={setOpenTrainModal}
+          />
         </div>
       </div>
       <UnlockModal isModalVisible={openUnlockModal} toggleIsVisible={toggleUnlockModal} />
